pyproject.toml
[tool.poetry]
name = "tuta"
version = "0.1.0"
description = ""
authors = ["Javier Javier <javierjdaza@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.1.4"
streamlit = "^1.29.0"
requests = "^2.31.0"
selenium = "^4.16.0"
bs4 = "^0.0.1"
webdriver-manager = "^4.0.1"
tqdm = "^4.66.1"
matplotlib = "^3.8.2"
plotly = "^5.18.0"
nbformat = "^5.9.2"
streamlit-option-menu = "^0.3.6"
supabase = "^2.3.4"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.28.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
